RecipeGenerator/requirements.txt
transformers
python-docx
fpdf
streamlit